tastr-infrastructure/src/main/java/eu/tastr/infrastructure/specification/OrSpecification.java
package eu.tastr.infrastructure.specification;

public class OrSpecification<T> implements Specification<T> {

    private final Specification<T> left;
    private final Specification<T> right;

    private OrSpecification(Specification<T> left, Specification<T> right) {
        this.left = left;
        this.right = right;
    }

    @Override
    public boolean isSatisfiedBy(T t) {
        return left.isSatisfiedBy(t) || right.isSatisfiedBy(t);
    }

    static <T> OrSpecification<T> or(Specification<T> left, Specification<T> right) {
        return new OrSpecification<>(left, right);
    }
}
